quick refactoring to externalize xml file creation
